# Support multi string contains [databricks]