Combine multi targets for multiple string contains, and invoke one kernel to get multiple bool columns results.

Depends on [cuDF PR](https://github.com/rapidsai/cudf/pull/16900)

Verified: `contains` in `case when` were combined when running pytest.

# Perf test

## summary
| strings pattern | total kernel time change | kernel speedup | end to end time change | end to end speedup |
|:-|:-|:-|:-|:-|
|short strings|125ms -> 38ms|3.5x|12s->12s|not obvious|
|long strings|24s -> 11s|2.18x|59s-53s|1.11x|

## details
### short strings
![image](https://github.com/user-attachments/assets/645691d1-3115-4363-b305-e5d690813f75)
The above part is kernel time of new kernel
The bottom part is base line.

### long strings
Kernel time:
Base line:
<img width="1069" alt="image" src="https://github.com/user-attachments/assets/4a08904a-4fc9-4b35-9fc5-85b12c240f42">

new:
<img width="1390" alt="image" src="https://github.com/user-attachments/assets/6ae6e10a-96c1-4540-b2c5-5e5eefbbdd0f">