@res-life 2024-11-18
The [previous commit](https://github.com/NVIDIA/spark-rapids/pull/11413/commits/5962eaaa8fd44a751e4758f65214005a925c3cd5) fixed the casting from String to UTF8String error:
```
E                   : java.lang.ClassCastException: java.lang.String cannot be cast to org.apache.spark.unsafe.types.UTF8String
E                   	at org.apache.spark.sql.rapids.ContainsCombiner.$anonfun$multiContains$3(stringFunctions.scala:506)
E                   	at scala.collection.TraversableLike.$anonfun$map$1(TraversableLike.scala:286)

```