@res-life 2024-11-18
Filed a follow-up issue [optimize the multi-contains generated by rlike](https://github.com/NVIDIA/spark-rapids/issues/11729)