@res-life 2024-11-18
The perf test in the description was updated, please look again.
For the long strings, kernel speedup is 2.18x now.
CI was passed, please review again.